+++ src/main/kotlin/com/group4/ticketingservice/dto/ReservationDto.kt
@@ -3,20 +3,20 @@ package com.group4.ticketingservice.dto
 import java.time.OffsetDateTime
 
 data class ReservationCreateRequest(
-    val eventId: Long
+    val eventId: Int
 )
 
 data class ReservationUpdateRequest(
-    val eventId: Long
+    val eventId: Int
 )
 
 data class ReservationDeleteRequest(
-    val id: Long
+    val id: Int
 )
 
 data class ReservationResponse(
-    val id: Long,
-    val eventId: Long,
-    val userId: Long,
+    val id: Int,
+    val eventId: Int,
+    val userId: Int,
     val bookedAt: OffsetDateTime
 )
